+++ test/unit/serialization/format_test.exs
@@ -1,0 +1,6 @@
+defmodule RDF.Serialization.FormatTest do
+  use ExUnit.Case
+
+  doctest RDF.Serialization.Format
+
+end